fix(loader): Loading paradigm Some philosophical changes of heart led to maintaining the API but relying more on the likely emerging changes in Razzle core.
Instead of overwriting the loader, plugin, this simply extends whatever
core does, which will be considered the canonical source. Further this only overwrites
the `.css` rule set. I noticed there is a module.css, but a PR can be
added to check for that and apply the same rules.

Also I added a basic test to confirm it works.